OREO-80 #comment Changed the variable averageRating type to double.

+++ app/src/main/java/com/oreo/mcommonjobs/Controllers/JobProviderRatingController.java
@@ -75,7 +75,7 @@ public class JobProviderRatingController {
                         int rating1 = ratingCurrentPosition.getInt("rating1");
                         int rating2 = ratingCurrentPosition.getInt("rating2");
                         int rating3 = ratingCurrentPosition.getInt("rating3");
-                        int averageRating = ratingCurrentPosition.getInt("averagerating");
+                        double averageRating = ratingCurrentPosition.getDouble("averagerating");
                         String comment = ratingCurrentPosition.getString("comment");
 
                         JobProviderRating jobProviderRating = new JobProviderRating(displayName, firstName, lastName, email, raterId, jobProviderId, rating1, rating2, rating3, averageRating, comment);
